django_app_docker/runapi/requirements.txt
amqp==5.0.6
asgiref==3.4.1
backcall==0.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==4.4.7
certifi==2021.5.30
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
colorlog==4.8.0
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.7
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
filetype==1.0.8
har2case==0.3.1
httpie==2.5.0
httpie-jwt-auth==0.4.0
HttpRunner==1.5.8
idna==3.2
importlib-metadata==4.8.1
inflection==0.5.1
ipython==7.16.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
jsonpath==0.82
kombu==5.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
mysqlclient==2.0.3
packaging==21.0
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.20
Pygments==2.10.0
PyJWT==1.7.1
pyparsing==2.4.7
PySocks==1.7.1
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-toolbelt==0.9.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
traitlets==4.3.3
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
zipp==3.5.0
